Breaking test dag into separate file for clarity's sake.
